\name{epicookbook-package}
\alias{epicookbook-package}
\alias{epicookbook}
\docType{package}
\title{
A cookbook of epidemiological models in R.
}
\description{
This package provides examples of different kinds of epidemiological models, and different methods for analyzing them, including simulation and parameter inference.
}
\details{
\tabular{ll}{
Package: \tab epicookbook\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-03-17\cr
License: \tab GPL (>= 2)\cr
}
Examples can be called either using \code{example} or \code{demo}.
}
\author{
Simon D.W. Frost <sdwfrost@gmail.com>
}
\references{
\itemize{
    \item{Anderson and May.}
  }
}
\keyword{ package }
\seealso{
\code{\link[simecol:simecol-package]{simecol}}
}
\examples{
demo(package="epicookbook")
}
